tests: restrict edit/delete metadata with storage provider only for synapse
