+++ tests/Backend/Workspaces/MetadataFromSnowflakeWorkspaceTest.php
@@ -4,15 +4,11 @@
 namespace Keboola\Test\Backend\Workspaces;
 
 use Keboola\Csv\CsvFile;
-use Keboola\StorageApi\ClientException;
 use Keboola\StorageApi\Metadata;
 use Keboola\Test\Backend\WorkspaceConnectionTrait;
-use Keboola\Test\Backend\Workspaces\Backend\WorkspaceBackendFactory;
 
 class MetadataFromSnowflakeWorkspaceTest extends ParallelWorkspacesTestCase
 {
-    const METADATA_PROVIDER_STORAGE = 'storage';
-
     use WorkspaceConnectionTrait;
 
     public function setUp()
@@ -506,76 +502,6 @@ class MetadataFromSnowflakeWorkspaceTest extends ParallelWorkspacesTestCase
         $this->assertEquals([], $table['columnMetadata']);
     }
 
-    public function testMetadataManipulationRestrictions()
-    {
-        // create workspace and source table in workspace
-        $workspace = $this->initTestWorkspace(self::BACKEND_SNOWFLAKE);
-
-        $db = $this->getDbConnection($workspace['connection']);
-        $db->query("create or replace table \"test.metadata_columns\" (
-                    \"string\" varchar(16) not null default 'string'
-                );");
-
-        // create table from workspace
-        $tableId = $this->_client->createTableAsyncDirect($this->getTestBucketId(self::STAGE_IN), array(
-            'name' => 'metadata_columns',
-            'dataWorkspaceId' => $workspace['id'],
-            'dataTableName' => 'test.metadata_columns',
-        ));
-
-        $expectedStringMetadata = [
-            'KBC.datatype.type' => 'TEXT',
-            'KBC.datatype.nullable' => '',
-            'KBC.datatype.basetype' => 'STRING',
-            'KBC.datatype.length' => '16',
-            'KBC.datatype.default' => '\'string\'',
-        ];
-
-        $metadata = new Metadata($this->_client);
-
-        $columnId = sprintf('%s.%s', $tableId, 'string');
-
-        $columnMetadataArray = $metadata->listColumnMetadata($columnId);
-        $this->assertCount(5, $columnMetadataArray);
-
-        $this->assertMetadata($expectedStringMetadata, $columnMetadataArray);
-
-        $columnMetadata = reset($columnMetadataArray);
-
-        // check that metadata from storage provider cannot be deleted
-        try {
-            $metadata->deleteColumnMetadata($columnId, $columnMetadata['id']);
-        } catch (ClientException $e) {
-            $this->assertSame(403, $e->getCode());
-            $this->assertSame('Metadata with "storage" provider cannot be deleted by user.', $e->getMessage());
-            $this->assertSame('storage.metadata.invalidProvider', $e->getStringCode());
-        }
-
-        // check that metadata with storage provider cannot be changed
-        try {
-            $metadata->postColumnMetadata(
-                $columnId,
-                self::METADATA_PROVIDER_STORAGE,
-                [
-                    [
-                        'key' => 'test',
-                        'value' => '1234',
-                    ],
-                    [
-                        'key' => $columnMetadata['key'],
-                        'value' => '1234',
-                    ],
-                ]
-            );
-        } catch (ClientException $e) {
-            $this->assertSame(403, $e->getCode());
-            $this->assertSame('Metadata with "storage" provider cannot be edited by user.', $e->getMessage());
-            $this->assertSame('storage.metadata.invalidProvider', $e->getStringCode());
-        }
-
-        $this->assertSame($columnMetadataArray, $metadata->listColumnMetadata($columnId));
-    }
-
     private function assertMetadata($expectedKeyValues, $metadata)
     {
         $this->assertEquals(count($expectedKeyValues), count($metadata));
@@ -586,7 +512,7 @@ class MetadataFromSnowflakeWorkspaceTest extends ParallelWorkspacesTestCase
             $this->assertArrayHasKey("provider", $data);
             $this->assertArrayHasKey("timestamp", $data);
             $this->assertRegExp(self::ISO8601_REGEXP, $data['timestamp']);
-            $this->assertEquals(self::METADATA_PROVIDER_STORAGE, $data['provider']);
+            $this->assertEquals(Metadata::PROVIDER_STORAGE, $data['provider']);
         }
     }
 }
